split up markets page from trading page
