+++ apps/trading-e2e/src/support/pages/markets-page.js
@@ -7,15 +7,6 @@ export default class MarketPage extends BasePage {
     'tradableInstrument.instrument.product.settlementAsset.symbol';
   marketRowPrices = 'flash-cell';
   marketRowDescription = 'name';
-  chartTab = 'chart';
-  ticketTab = 'ticket';
-  orderbookTab = 'orderbook';
-  ordersTab = 'orders';
-  positionsTab = 'positions';
-  collateralTab = 'collateral';
-  tradesTab = 'trades';
-  completedTrades = 'market-trades';
-  orderBookTab = 'orderbook';
 
   validateMarketTableDisplayed() {
     const expectedMarketHeaders = [
@@ -58,10 +49,6 @@ export default class MarketPage extends BasePage {
     );
   }
 
-  validateCompletedTradesDisplayed() {
-    cy.getByTestId(this.completedTrades).should('not.be.empty');
-  }
-
   clickOnMarket(marketText) {
     cy.contains(marketText).click();
   }
@@ -72,25 +59,5 @@ export default class MarketPage extends BasePage {
 
   clickOnTopMarketRow() {
     cy.get('[col-id="data"]').eq(1).click();
-  }
-
-  clickOnOrdersTab() {
-    cy.getByTestId(this.ordersTab).click();
-  }
-
-  clickOnTicketTab() {
-    cy.getByTestId(this.ticketTab).click();
-  }
-
-  clickOnCollateralTab() {
-    cy.getByTestId(this.collateralTab).click();
-  }
-
-  clickOnTradesTab() {
-    cy.getByTestId(this.tradesTab).click();
-  }
-
-  clickOrderBookTab() {
-    cy.getByTestId(this.orderBookTab).click();
   }
 }
