+++ apps/trading-e2e/src/support/step_definitions/common-step.js
@@ -1,5 +1,5 @@
-import { Given, When } from 'cypress-cucumber-preprocessor/steps';
-import MarketsPage from '../pages/markets-page';
+import { Given } from 'cypress-cucumber-preprocessor/steps';
+import MarketsPage from '../pages/trading-page';
 
 const marketsPage = new MarketsPage();
